Rename packages Fix missing service loader
