--- common/src/main/java/com/example/examplemod/Constants.java
@@ -1,11 +1,0 @@
-package com.example.examplemod;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class Constants {
-
-	public static final String MOD_ID = "examplemod";
-	public static final String MOD_NAME = "ExampleMod";
-	public static final Logger LOG = LoggerFactory.getLogger(MOD_NAME);
-}
