+++ common/src/main/java/de/macbrayne/common/inventorypause/Constants.java
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: EUPL-1.2
+
+package de.macbrayne.common.inventorypause;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class Constants {
+
+	public static final String MOD_ID = "inventorypause";
+	public static final String MOD_NAME = "InventoryPause";
+	public static final Logger LOG = LoggerFactory.getLogger(MOD_NAME);
+}
