+++ common/src/main/java/de/macbrayne/common/inventorypause/mixin/MixinMinecraft.java
@@ -1,20 +1,21 @@
-package com.example.examplemod.mixin;
+// SPDX-License-Identifier: EUPL-1.2
 
-import com.example.examplemod.Constants;
+package de.macbrayne.common.inventorypause.mixin;
+
+import de.macbrayne.common.inventorypause.Constants;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.screens.TitleScreen;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(TitleScreen.class)
-public class MixinTitleScreen {
+@Mixin(Minecraft.class)
+public class MixinMinecraft {
     
-    @Inject(at = @At("HEAD"), method = "init()V")
+    @Inject(at = @At("TAIL"), method = "<init>")
     private void init(CallbackInfo info) {
         
-        Constants.LOG.info("This line is printed by an example mod mixin from Fabric!");
+        Constants.LOG.info("This line is printed by an example mod common mixin!");
         Constants.LOG.info("MC Version: {}", Minecraft.getInstance().getVersionType());
     }
 }
