+++ common/src/main/java/de/macbrayne/common/inventorypause/platform/Services.java
@@ -1,7 +1,9 @@
-package com.example.examplemod.platform;
+// SPDX-License-Identifier: EUPL-1.2
 
-import com.example.examplemod.Constants;
-import com.example.examplemod.platform.services.IPlatformHelper;
+package de.macbrayne.common.inventorypause.platform;
+
+import de.macbrayne.common.inventorypause.Constants;
+import de.macbrayne.common.inventorypause.platform.services.IPlatformHelper;
 
 import java.util.ServiceLoader;
