+++ fabric/src/main/java/de/macbrayne/fabric/inventorypause/mixin/MixinTitleScreen.java
@@ -1,19 +1,22 @@
-package com.example.examplemod.mixin;
+// SPDX-License-Identifier: EUPL-1.2
 
-import com.example.examplemod.Constants;
+package de.macbrayne.fabric.inventorypause.mixin;
+
+import de.macbrayne.common.inventorypause.Constants;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.screens.TitleScreen;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(Minecraft.class)
-public class MixinMinecraft {
+@Mixin(TitleScreen.class)
+public class MixinTitleScreen {
     
-    @Inject(at = @At("TAIL"), method = "<init>")
+    @Inject(at = @At("HEAD"), method = "init()V")
     private void init(CallbackInfo info) {
         
-        Constants.LOG.info("This line is printed by an example mod common mixin!");
+        Constants.LOG.info("This line is printed by an example mod mixin from Fabric!");
         Constants.LOG.info("MC Version: {}", Minecraft.getInstance().getVersionType());
     }
 }
