+++ fabric/src/main/resources/META-INF/services/de.macbrayne.common.inventorypause.platform.services.IPlatformHelper
@@ -1,0 +1,1 @@
+de.macbrayne.fabric.inventorypause.platform.FabricPlatformHelper
