+++ neoforge/src/main/java/de/macbrayne/forge/inventorypause/platform/NeoForgePlatformHelper.java
@@ -1,0 +1,27 @@
+// SPDX-License-Identifier: EUPL-1.2
+
+package de.macbrayne.forge.inventorypause.platform;
+
+import de.macbrayne.common.inventorypause.platform.services.IPlatformHelper;
+import net.neoforged.fml.ModList;
+import net.neoforged.fml.loading.FMLLoader;
+
+public class NeoForgePlatformHelper implements IPlatformHelper {
+    @Override
+    public String getPlatformName() {
+
+        return "NeoForge";
+    }
+
+    @Override
+    public boolean isModLoaded(String modId) {
+
+        return ModList.get().isLoaded(modId);
+    }
+
+    @Override
+    public boolean isDevelopmentEnvironment() {
+
+        return !FMLLoader.isProduction();
+    }
+}
